% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushfire_risk-data.R
\docType{data}
\name{bushfire_risk}
\alias{bushfire_risk}
\title{Bushfire risk summary data}
\format{
A tibble with 4 rows and 7 variables:
}
\usage{
data(bushfire_risk)
}
\description{
A small summary dataset derived from the Black Summer bushfire
attribution analysis.
}
\keyword{datasets}
