% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bushfire_app.R
\name{run_bushfire_app}
\alias{run_bushfire_app}
\title{Run the Black Summer Interactive Shiny App}
\usage{
run_bushfire_app()
}
\value{
Opens the bushfire risk explorer Shiny app.
}
\description{
Run the Black Summer Interactive Shiny App
}
